bugfix: add checking whether info->bus is empty
When the dbus service is restarted quickly and frequently, the
Dbus bus will not be found in a short period of time. As a result,
the info->bus obtained by the virDbusWatchCallback() for monitoring the
dbus in libvirt is empty, and a core dump occurs in the libvirt.
Corresponding judgment is added to prevent this core dump.

Signed-off-by: mayunlong <mayunlong6@huawei.com>