conf: Introduce SEV-SNP support
commit c65eba1f57ab1670f5bfc6bd178df1d5c3b09e2a upstream

SEV-SNP is an enhancement of SEV/SEV-ES and thus it shares some
fields with it. Nevertheless, on XML level, it's yet another type
of <launchSecurity/>.

[Backport Changes]

1. The enumeration value VIR_DOMAIN_LAUNCH_SECURITY_CVM, which is
a member of the virDomainLaunchSecurity enum, was introduced in
Euler libvirt as part of CSV support in commit ddf9053. However,
the switch statements in the current patch do not account for this
new enum member. This backport ensures that
VIR_DOMAIN_LAUNCH_SECURITY_CVM is correctly handled by adding the
appropriate case block in the function
virDomainDefLaunchSecurityValidate() in the file
src/conf/domain_validate.c. This change prevents compiler warnings
and enhances the completeness of the code.

2. The addition of the launch-security-sev-snp test
cases(in tests/qemuxml2argvdata/) to the Euler Libvirt 9.10 test
framework was skipped because it requires complete backport of the
latest test framework from Libvirt 10.5 and above. So,this commit
prioritizes enabling SEV-SNP in Libvirt 9.10 and excludes the changes
to the files listed below :
tests/qemuxml2argvdata/launch-security-sev-snp.x86_64-latest.args
tests/qemuxml2argvdata/launch-security-sev-snp.x86_64-latest.xml
tests/qemuxml2argvdata/launch-security-sev-snp.xml

3.In src/qemu/qemu_capabilities.c, within the function
virQEMUCapsFillDomainLaunchSecurity(), an if condition was added that
enables support for QEMU_CAPS_SEV_SNP_GUEST capability. This change comes
from a different upstream commit 66df799 which was already backported
into OpenEuler upstream branch but excluded the aforementioned change.
Therefore, QEMU_CAPS_SEV_SNP_GUEST capability support has been enabled
in the current backport commit.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Reviewed-by: Daniel P. Berrangé <berrange@redhat.com>
Signed-off-by: PrithivishS <sprithiv@amd.com>