src: Convert some _virDomainSecDef::sectype checks to switch()
commit 7d16c296e3f0ed61443d3095fd6f8951de57f79d upstream

In a few instances there is a plain if() check for
_virDomainSecDef::sectype. While this works perfectly for now,
soon there'll be another type and we can utilize compiler to
identify all the places that need adaptation. Switch those if()
statements to switch().

[Backport Changes]

1.The enumeration value VIR_DOMAIN_LAUNCH_SECURITY_CVM, a member of the
virDomainLaunchSecurity enum, was introduced in libvirt as part of
CSV support in commit ddf9053ad7d. However,the switch statements in the
upstream patch do not account for this new enum member. This backport
ensures that VIR_DOMAIN_LAUNCH_SECURITY_CVM is correctly handled in all
the newly added switch statements by adding the appropriate case block in
all affected files. This change prevents compiler warnings and enhances
the completeness of the code.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Reviewed-by: Daniel P. Berrangé <berrange@redhat.com>
Signed-off-by: TauqirAzam <MdTauqir.AzamKausar@amd.com>
Signed-off-by: PrithivishS <sprithiv@amd.com>