migration: Add the VIR_MIGRATE_RETURNPATH flag
In the past we turned on the migration feature QEMU_MIGRATION_CAP_RETURN_PATH
by default, which accelerated the recovery of vcpu during the hot migration rollback,
but at the same time increased the downtime of the hot migration. so we made it easier
by adding a new libvirt parameter (return-path) as a switch for the QEMU_ MIGRATION_CAP_RETURN_PATH
feature, while removing the previous VIR_MIGRATE_TUNNELLED related control logic for return-path.
Signed-off-by: yangtao <2429917001@qq.com>